apps/api/requirements.txt
fastapi>=0.120.0
uvicorn>=0.38.0
asyncpg>=0.30.0
psycopg>=3.2.0
pyjwt[crypto]>=2.10.0
python-dotenv>=1.1.1
